Major refactor: MySQL + gorm, grpc-gateway
This commit includes a few weeks of misc. work moving this project
off of DynamoDB and onto MySQL to use more standard tooling. We'll be
using Aurora Serverless in production. This now uses grpc-gateway to
offer an HTTP endpoint to access the various gRPC endpoints in order
to make it easier to test backend endpoints.